==> 06-interactive.Rmd <==
# Interactive component

In the last section, we analyzed payment choices over a variety of factors using data in 2019. Now, we are interested in how these correlations change over time. We will focus on the correlation between age and payment choices. In this section, we utilize D3 to build an interactive stacked bar plot. Each plot shows the change of payment choices in a certain age group. You can click on the button at the upper right corner to switch to different age groups. You can also hover over the colored blocks in the graph to highlight that specific payment type. This interactive presentation is more clear and convenient to view than facets.

```{r include=FALSE}
htmltools::tagList(rmarkdown::html_dependency_font_awesome())
```

<script src="https://d3js.org/d3.v6.min.js"></script>
<script src="https://cdnjs.cloudflare.com/ajax/libs/d3-legend/2.25.6/d3-legend.min.js"></script>
<p>Changes of payment choice over time in the age group: <select id="selectButton"></select></p>
<div id="my_dataviz"></div>
<script src="interactive.js"></script>

We can see that from 2015 to 2019, paper payments (cash, checks and money orders) show a overall decline among most groups. Particularly, cash payments became less popular in the group of 18-25. In the year of 2016, card payments popularity had a noticeable increase and stayed at that level for the next three years.

In last section, we used data from 2019. In this interactive plot, we can see that the difference among age groups stay the same over the 5 years. Younger people tend to like paper payments more while the elders prefer cards and online payments. 


Citation: 

1. https://www.d3-graph-gallery.com/graph/barplot_stacked_percent.html

2. https://www.d3-graph-gallery.com/graph/barplot_stacked_highlight.html
