==> 07-conclusion.Rmd <==
# Conclusion

We analyzed the effect of individual factors and transaction situations on payment choices. It is exciting that payment preferences can be influenced by so many features. However, due to the unbalanced data, it is hard to analyze the influence of certain factors over the payment choices at the exploratory level. For example, in the stacked bar plots, most payment types are hard to see due to the limited records. For improvements, we can collect more data or build separate graphs for the payment types with limited records. Through this project, we learned that graph choice is crucial. When a graph has no obvious information, another kind of plot will have a stunning different result. Also, when a graph shows no obvious correlation, we can still find useful information from it.
